Add CI/CD via github actions for the NewPaletteSystem Branch Builds for ubuntu x64 using gcc and clang Builds for windows x64 using MSVC Test are ran on all platforms Some currently fail on windows beacuse the lua51.dll is missing in the directory where test .exe is located

+++ src/Mobs/AggressiveMonster.cpp
@@ -91,7 +91,7 @@ cMonster * cAggressiveMonster::GetMonsterOfTypeInSight(eEntityType a_MobType, un
 		return false;
 	};
 
-	cBoundingBox CheckZone(GetPosition().addedXZ(-a_SightDistance, -a_SightDistance), GetPosition().addedXZ(a_SightDistance, a_SightDistance));
+	cBoundingBox CheckZone(GetPosition().addedXZ(-static_cast<int>(a_SightDistance), -static_cast<int>(a_SightDistance)), GetPosition().addedXZ(a_SightDistance, a_SightDistance));
 	m_World->ForEachEntityInBox(CheckZone, Callback);
 	return FoundTarget;
 }
